refactor(ReservationTimeService): 이미 예약이 되어 있는 시간은 삭제없는 걸 검증하는 로직 추가

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -44,6 +44,10 @@ public class ReservationTimeService {
 
     @Transactional
     public void delete(Long id) {
+        if(reservationRepository.existByTimeId(id)) {
+            throw new DomainException(ErrorCode.RESERVATION_TIME_HAS_RESERVATION);
+        }
+
         reservationTimeRepository.deleteById(id);
     }
